fix(agent): rename session fork module
